added formatting to all files
Also added
Theme
Category skeleton page
About/SystemInfo page
