+++ Firefly/FireflyAPIClient/Errors/CommonError.swift
@@ -8,20 +8,20 @@
 import Foundation
 
 struct CommonError: Codable {
-    let message: String?
-    let exception: String?
+  let message: String?
+  let exception: String?
 }
 
 enum CommonErrorThrow: Error {
-    case decodeError
+  case decodeError
 }
 
 func ErrorDecoder(_ errorData: Data) throws -> CommonError {
-    do {
-        let decoder = JSONDecoder()
-        let errorResult = try decoder.decode(CommonError.self, from: errorData)
-        return errorResult
-    } catch {
-        throw CommonErrorThrow.decodeError
-    }
+  do {
+    let decoder = JSONDecoder()
+    let errorResult = try decoder.decode(CommonError.self, from: errorData)
+    return errorResult
+  } catch {
+    throw CommonErrorThrow.decodeError
+  }
 }
