+++ Firefly/FireflyApp.swift
@@ -13,6 +13,7 @@ struct FireflyApp: App {
   @StateObject private var alertViewModel = AlertViewModel()
   @StateObject private var menuViewModel = MenuViewModel()
   @Environment(\.scenePhase) var scenePhase
+  @AppStorage("selectedTheme") private var selectedTheme: ThemeMode = .system
 
   init() {
     #if DEBUG
@@ -28,10 +29,15 @@ struct FireflyApp: App {
   var body: some Scene {
     WindowGroup {
       Menu()
-        .preferredColorScheme(.dark)
+        .preferredColorScheme(colorScheme)
         .environmentObject(alertViewModel)
         .environmentObject(menuViewModel)
         .toast(isPresenting: $alertViewModel.show, alert: { alertViewModel.alertToast })
+        .onOpenURL { url in
+          if url.scheme == "shortcut" && url.host == "addTransaction" {
+            menuViewModel.openTransactionSheet()
+          }
+        }
     }.onChange(of: scenePhase) { _, newPhase in
       switch newPhase {
       case .active:
@@ -47,11 +53,25 @@ struct FireflyApp: App {
 
     }
   }
+  private var colorScheme: ColorScheme? {
+    switch selectedTheme {
+    case .light:
+      return .light
+    case .dark:
+      return .dark
+    case .system:
+      return nil
+    }
+  }
 }
 
 func updateShortcutItems() {
   let addTransactionShortcut = UIApplicationShortcutItem(
-    type: "addTransaction", localizedTitle: "Add Transaction", localizedSubtitle: "",
-    icon: UIApplicationShortcutIcon(systemImageName: "plus.app.fill"), userInfo: [:])
+    type: "addTransaction",
+    localizedTitle: "Add Transaction",
+    localizedSubtitle: "",
+    icon: UIApplicationShortcutIcon(systemImageName: "plus.app.fill"),
+    userInfo: [:]
+  )
   UIApplication.shared.shortcutItems = [addTransactionShortcut]
 }
