+++ Firefly/Helpers/BackgroundMaterial.swift
@@ -24,30 +24,30 @@ private let br = Color("bottom_right")
 // MARK: - Grainy gradient
 
 extension ShapeStyle where Self == AnyShapeStyle {
-    static func grainGradient(time: TimeInterval, gridSize: Int = 3) -> Self {
-        return AnyShapeStyle(
-            ShaderLibrary.default.grainGradient(
-                .boundingRect,
-                .float(3),
-                .float(time),
-                .colorArray([
-                    tl, tc, tr,
-                    ml, mc, mr,
-                    bl, bc, br,
-                ])
-            ))
-    }
+  static func grainGradient(time: TimeInterval, gridSize: Int = 3) -> Self {
+    return AnyShapeStyle(
+      ShaderLibrary.default.grainGradient(
+        .boundingRect,
+        .float(3),
+        .float(time),
+        .colorArray([
+          tl, tc, tr,
+          ml, mc, mr,
+          bl, bc, br,
+        ])
+      ))
+  }
 }
 
 // MARK: - Leather
 
 extension ShapeStyle where Self == AnyShapeStyle {
-    static func leather(lightColor: Color, time: TimeInterval) -> Self {
-        return AnyShapeStyle(
-            ShaderLibrary.default.leather(
-                .boundingRect,
-                .color(lightColor),
-                .float(time)
-            ))
-    }
+  static func leather(lightColor: Color, time: TimeInterval) -> Self {
+    return AnyShapeStyle(
+      ShaderLibrary.default.leather(
+        .boundingRect,
+        .color(lightColor),
+        .float(time)
+      ))
+  }
 }
