+++ Firefly/Helpers/DebugHelpers.swift
@@ -8,9 +8,9 @@
 import Foundation
 
 func printResponse(_ data: Data) {
-    if let responseString = String(data: data, encoding: .utf8) {
-        print(responseString)
-    } else {
-        print("Unable to convert data")
-    }
+  if let responseString = String(data: data, encoding: .utf8) {
+    print(responseString)
+  } else {
+    print("Unable to convert data")
+  }
 }
