+++ Firefly/Helpers/Loading.swift
@@ -8,17 +8,17 @@
 import SwiftUI
 
 struct LoadingSpinner: View {
-    var body: some View {
-        Spacer()
-        HStack(alignment: .center) {
-            Spacer()
-            ProgressView()
-            Spacer()
-        }
-        Spacer()
+  var body: some View {
+    Spacer()
+    HStack(alignment: .center) {
+      Spacer()
+      ProgressView()
+      Spacer()
     }
+    Spacer()
+  }
 }
 
 #Preview {
-    LoadingSpinner()
+  LoadingSpinner()
 }
