+++ Firefly/Resources/Localizable.xcstrings
@@ -1,0 +1,402 @@
+{
+  "sourceLanguage" : "en",
+  "strings" : {
+    "" : {
+
+    },
+    "About" : {
+
+    },
+    "About Firefly" : {
+
+    },
+    "Accounts" : {
+
+    },
+    "Add" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "追加"
+          }
+        }
+      }
+    },
+    "Add Category" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "カテゴリ追加"
+          }
+        }
+      }
+    },
+    "Add Transaction" : {
+
+    },
+    "Amount" : {
+
+    },
+    "Amount spend" : {
+
+    },
+    "API Version: %@" : {
+
+    },
+    "Appearance" : {
+
+    },
+    "Apply Filter" : {
+
+    },
+    "Are you sure you want to delete this transaction?" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "削除してもよろしいでしょうか？"
+          }
+        }
+      }
+    },
+    "As of %@" : {
+
+    },
+    "Asset Account" : {
+
+    },
+    "Budget" : {
+
+    },
+    "Cancel" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "キャンセル"
+          }
+        }
+      }
+    },
+    "Categories" : {
+      "extractionState" : "manual",
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Categories"
+          }
+        },
+        "ja" : {
+          "stringUnit" : {
+            "state" : "needs_review",
+            "value" : "カテゴリ"
+          }
+        }
+      }
+    },
+    "Category" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "カテゴリ"
+          }
+        }
+      }
+    },
+    "Category Summary" : {
+
+    },
+    "Credentials" : {
+
+    },
+    "Current balance" : {
+
+    },
+    "Current Month" : {
+
+    },
+    "Date" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "日付"
+          }
+        }
+      }
+    },
+    "DEBUG" : {
+
+    },
+    "Delete" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "削除"
+          }
+        }
+      }
+    },
+    "Description" : {
+
+    },
+    "Destination" : {
+
+    },
+    "Details" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "詳細"
+          }
+        }
+      }
+    },
+    "Done" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "OK"
+          }
+        }
+      }
+    },
+    "Driver: %@" : {
+
+    },
+    "Edit" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "編集"
+          }
+        }
+      }
+    },
+    "Edit Transaction" : {
+
+    },
+    "End Date" : {
+
+    },
+    "Enter keywords in the search bar above" : {
+
+    },
+    "Enter the name for a new category" : {
+
+    },
+    "Error" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "エラー"
+          }
+        }
+      }
+    },
+    "Filter" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "絞り込み"
+          }
+        }
+      }
+    },
+    "Firefly Client Version: %@" : {
+
+    },
+    "Group Title" : {
+
+    },
+    "Home" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "ホーム"
+          }
+        }
+      }
+    },
+    "Instructions" : {
+
+    },
+    "Loaded" : {
+
+    },
+    "More Info" : {
+
+    },
+    "New category" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "カテゴリ追加"
+          }
+        }
+      }
+    },
+    "No data available" : {
+
+    },
+    "No results found" : {
+
+    },
+    "Notes" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "メモ"
+          }
+        }
+      }
+    },
+    "OS: %@" : {
+
+    },
+    "PHP Version: %@" : {
+
+    },
+    "Please enter a name for the new category." : {
+
+    },
+    "Recent Transactions" : {
+
+    },
+    "Reset" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "リセット"
+          }
+        }
+      }
+    },
+    "Results for \"%@\"" : {
+
+    },
+    "Save" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "保存"
+          }
+        }
+      }
+    },
+    "Search" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "検索"
+          }
+        }
+      }
+    },
+    "Search for transactions" : {
+
+    },
+    "Search..." : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "検索"
+          }
+        }
+      }
+    },
+    "Select an account" : {
+
+    },
+    "Select an item" : {
+
+    },
+    "Settings" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "設定"
+          }
+        }
+      }
+    },
+    "Something went wrong" : {
+
+    },
+    "Something went wrong." : {
+
+    },
+    "Source" : {
+
+    },
+    "Split #%lld" : {
+
+    },
+    "Start Date" : {
+
+    },
+    "Suggestions" : {
+
+    },
+    "Theme" : {
+
+    },
+    "Total Amount" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "会計"
+          }
+        }
+      }
+    },
+    "Transaction Details" : {
+
+    },
+    "Transaction ID: %@" : {
+
+    },
+    "Transaction Options" : {
+
+    },
+    "Transactions" : {
+
+    },
+    "Try Again" : {
+
+    },
+    "Type" : {
+
+    },
+    "Uncategorized" : {
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "カテゴリなし"
+          }
+        }
+      }
+    }
+  },
+  "version" : "1.0"
+}
