+++ Firefly/Views/Settings/Categories/CategoriesView.swift
@@ -1,0 +1,22 @@
+//
+//  Categories.swift
+//  Firefly
+//
+//  Created by Aditya Srinivasa on 2025/06/05.
+//
+
+import SwiftUI
+
+struct CategoriesView: View {
+  var body: some View {
+    NavigationStack {
+
+    }
+    .navigationTitle("Categories")
+    .navigationBarTitleDisplayMode(.large)
+  }
+}
+
+#Preview {
+  CategoriesView()
+}
